Include Username in routes
The username has been added as a Header when using the exploration service so that an explored planet can be linked to a user.

Now you can retrieve all the planets explored by a user. There's a minor error on the explore route, that doesn't wait until ollama finishes generating and makes the gateway return server error